requirements.txt
python-dotenv==0.20.0
aiogram==2.25.2
SQLAlchemy==1.3.24
gino==1.0.1
asyncpg==0.26.0

